Changes in player input and Invalid input.
Made player input more accurate. For example: befor it was Choose [r]ock and now it is [r] for rock. This way is more understadable.
next change is for invalid player input. If player chose letter differnt from "r, s or p" it will appear a certain massage and the loop will continue.